move rename util
