+++ python/datasources/phrma.py
@@ -1,5 +1,5 @@
 import pandas as pd
-from typing import Dict, cast
+from typing import cast
 from datasources.data_source import DataSource
 from ingestion.constants import (
     COUNTY_LEVEL,
@@ -23,7 +23,28 @@ from ingestion.het_types import (
     PHRMA_BREAKDOWN_TYPE_OR_ALL,
     PHRMA_BREAKDOWN_TYPE,
 )
-from ingestion.phrma_utils import TMP_ALL, PHRMA_DIR, get_sheet_name
+from ingestion.phrma_utils import (
+    TMP_ALL,
+    PHRMA_DIR,
+    get_sheet_name,
+    ADHERENCE_RATE,
+    PER_100K,
+    MEDICARE_DISEASE_COUNT,
+    COUNT_TOTAL,
+    COUNT_YES,
+    RACE_NAME,
+    AGE_GROUP,
+    MEDICARE_POP_COUNT,
+    SEX_NAME,
+    LIS,
+    ENTLMT_RSN_CURR,
+    STATE_FIPS,
+    COUNTY_FIPS,
+    PHRMA_PCT_CONDITIONS,
+    PHRMA_100K_CONDITIONS,
+    rename_cols,
+)
+
 
 """
 NOTE: Phrma data comes in .xlsx files, with breakdowns by sheet.
@@ -38,41 +59,6 @@ ELIGIBILITY = "eligibility"
 ADHERENCE = 'adherence'
 BENEFICIARIES = 'beneficiaries'
 DTYPE = {'COUNTY_FIPS': str, 'STATE_FIPS': str}
-
-PHRMA_PCT_CONDITIONS = [
-    std_col.ARV_PREFIX,
-    std_col.BETA_BLOCKERS_PREFIX,
-    std_col.CCB_PREFIX,
-    std_col.DOAC_PREFIX,
-    std_col.BB_AMI_PREFIX,
-    std_col.RASA_PREFIX,
-    std_col.STATINS_PREFIX,
-    std_col.ANTI_PSYCHOTICS_PREFIX,
-]
-
-PHRMA_100K_CONDITIONS = [
-    std_col.AMI_PREFIX,
-    std_col.PHRMA_HIV_PREFIX,
-    std_col.SCHIZOPHRENIA_PREFIX,
-]
-
-# CONSTANTS USED BY DATA SOURCE
-COUNT_TOTAL = "TOTAL_BENE"
-COUNT_YES = "BENE_YES"
-# COUNT_NO = "BENE_NO"
-MEDICARE_DISEASE_COUNT = "BENE_N"
-MEDICARE_POP_COUNT = "TOTAL_N"
-ADHERENCE_RATE = "BENE_YES_PCT"
-PER_100K = "PER_100K"
-STATE_NAME = "STATE_NAME"
-COUNTY_NAME = "COUNTY_NAME"
-STATE_FIPS = "STATE_FIPS"
-COUNTY_FIPS = "COUNTY_FIPS"
-ENTLMT_RSN_CURR = "ENTLMT_RSN_CURR"
-LIS = "LIS"
-RACE_NAME = "RACE_NAME"
-AGE_GROUP = "AGE_GROUP"
-SEX_NAME = "SEX_NAME"
 
 
 # a nested dictionary that contains values swaps per column name
@@ -346,54 +332,3 @@ def load_phrma_df_from_data_dir(geo_level: GEO_TYPE, breakdown: PHRMA_BREAKDOWN_
     df_merged = ensure_leading_zeros(df_merged, fips_col, fips_length)
 
     return df_merged
-
-
-def rename_cols(
-    df: pd.DataFrame,
-    geo_level: GEO_TYPE,
-    breakdown: PHRMA_BREAKDOWN_TYPE_OR_ALL,
-    condition: str,
-) -> pd.DataFrame:
-    """Renames columns based on the demo/geo breakdown"""
-
-    rename_cols_map: Dict[str, str] = (
-        {
-            COUNT_YES: f'{condition}_{COUNT_YES}',
-            COUNT_TOTAL: f'{condition}_{COUNT_TOTAL}',
-            ADHERENCE_RATE: f'{condition}_{ADHERENCE_RATE}',
-        }
-        if condition in PHRMA_PCT_CONDITIONS
-        else {
-            MEDICARE_DISEASE_COUNT: f'{condition}_{MEDICARE_DISEASE_COUNT}',
-            PER_100K: f'{condition}_{PER_100K}',
-        }
-    )
-
-    if geo_level == COUNTY_LEVEL:
-        rename_cols_map[COUNTY_FIPS] = std_col.COUNTY_FIPS_COL
-
-    if geo_level in [STATE_LEVEL, NATIONAL_LEVEL]:
-        rename_cols_map[STATE_FIPS] = std_col.STATE_FIPS_COL
-
-    if breakdown == std_col.RACE_OR_HISPANIC_COL:
-        rename_cols_map[RACE_NAME] = std_col.RACE_CATEGORY_ID_COL
-
-    if breakdown == std_col.SEX_COL:
-        rename_cols_map[SEX_NAME] = std_col.SEX_COL
-
-    if breakdown == std_col.AGE_COL:
-        rename_cols_map[AGE_GROUP] = std_col.AGE_COL
-
-    if breakdown == std_col.ELIGIBILITY_COL:
-        rename_cols_map[ENTLMT_RSN_CURR] = std_col.ELIGIBILITY_COL
-
-    if breakdown == std_col.LIS_COL:
-        rename_cols_map[LIS] = std_col.LIS_COL
-
-    df = df.rename(columns=rename_cols_map)
-
-    # only keep the medicare/medicaid raw population for one of the 100k conditions
-    if condition in PHRMA_100K_CONDITIONS and condition != std_col.AMI_PREFIX:
-        df = df.drop(columns=[MEDICARE_POP_COUNT])
-
-    return df
